separate out config (for site-specific deployment)
